Introduce templates
Themes can make use of templates, allows to change visuals faster and
without copying everything, also all themes make use of a template are
profiting from bug fixes and enhancements of this template. Less
error-pone, more robust.

Templates might be also the foundation in the future for full user
customization of themes.

For the beginning I ported the Dark theme to it.