Rebase the Darwin theme on the Basic template
This is required to make it work again with the WebExtension version.
In order to rebase it I have advanced the abilities of the Basic template.
The Dark and Light themes are using this abilities as well now.

I also did a a few minor other improvements. For example, this improves the layout in Opera.